Moved DataReflectionVisitor to a header file
